REFACTOR: Presentation 레이어 내 유틸 및 확장 함수 파일 경로 수정, 일부 미사용 파일 제거
