--- presentation/src/main/java/com/threegap/bitnagil/presentation/common/flow/CollectAsEffect.kt
@@ -1,22 +1,0 @@
-package com.threegap.bitnagil.presentation.common.flow
-
-import android.annotation.SuppressLint
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
-import kotlin.coroutines.CoroutineContext
-import kotlin.coroutines.EmptyCoroutineContext
-
-@SuppressLint("ComposableNaming")
-@Composable
-fun <T> Flow<T>.collectAsEffect(
-    context: CoroutineContext = EmptyCoroutineContext,
-    block: (T) -> Unit,
-) {
-    LaunchedEffect(Unit) {
-        onEach(block).flowOn(context).launchIn(this)
-    }
-}
