+++ presentation/src/main/java/com/threegap/bitnagil/presentation/screen/recommendroutine/RecommendRoutineScreen.kt
@@ -32,9 +32,9 @@ import com.threegap.bitnagil.designsystem.component.atom.BitnagilIcon
 import com.threegap.bitnagil.designsystem.component.block.BitnagilTopBar
 import com.threegap.bitnagil.designsystem.modifier.clickableWithoutRipple
 import com.threegap.bitnagil.domain.recommendroutine.model.RecommendCategory
-import com.threegap.bitnagil.presentation.common.extension.displayLevel
-import com.threegap.bitnagil.presentation.common.extension.displayTitle
-import com.threegap.bitnagil.presentation.common.extension.isVisible
+import com.threegap.bitnagil.presentation.model.routine.displayLevel
+import com.threegap.bitnagil.presentation.model.routine.displayTitle
+import com.threegap.bitnagil.presentation.model.routine.isVisible
 import com.threegap.bitnagil.presentation.screen.recommendroutine.component.atom.RecommendCategoryChip
 import com.threegap.bitnagil.presentation.screen.recommendroutine.component.block.EmotionRecommendRoutineButton
 import com.threegap.bitnagil.presentation.screen.recommendroutine.component.block.RecommendRoutineItem
