+++ presentation/src/main/java/com/threegap/bitnagil/presentation/screen/recommendroutine/component/template/RecommendLevelBottomSheet.kt
@@ -19,8 +19,8 @@ import com.threegap.bitnagil.designsystem.R
 import com.threegap.bitnagil.designsystem.component.atom.BitnagilIcon
 import com.threegap.bitnagil.designsystem.modifier.clickableWithoutRipple
 import com.threegap.bitnagil.domain.recommendroutine.model.RecommendLevel
-import com.threegap.bitnagil.presentation.common.extension.displayLevel
-import com.threegap.bitnagil.presentation.common.extension.displayTitle
+import com.threegap.bitnagil.presentation.model.routine.displayLevel
+import com.threegap.bitnagil.presentation.model.routine.displayTitle
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
