+++ presentation/src/main/java/com/threegap/bitnagil/presentation/util/dimension/DpToPx.kt
@@ -1,4 +1,4 @@
-package com.threegap.bitnagil.presentation.common.dimension
+package com.threegap.bitnagil.presentation.util.dimension
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalDensity
