+++ presentation/src/main/java/com/threegap/bitnagil/presentation/util/playstore/PlayStoreUtils.kt
@@ -1,4 +1,4 @@
-package com.threegap.bitnagil.presentation.common.playstore
+package com.threegap.bitnagil.presentation.util.playstore
 
 import android.content.ActivityNotFoundException
 import android.content.Context
