+++ presentation/src/main/java/com/threegap/bitnagil/presentation/util/playstore/UpdateAvailableState.kt
@@ -1,4 +1,4 @@
-package com.threegap.bitnagil.presentation.common.playstore
+package com.threegap.bitnagil.presentation.util.playstore
 
 enum class UpdateAvailableState {
     LATEST, NEED_UPDATE, NONE
