+++ presentation/src/main/java/com/threegap/bitnagil/presentation/util/version/VersionNameProvider.kt
@@ -1,4 +1,4 @@
-package com.threegap.bitnagil.presentation.common.version
+package com.threegap.bitnagil.presentation.util.version
 
 interface VersionNameProvider {
     fun getVersionName(): String
